Изменен тип данных в task1, начато решение задачи task2

+++ Tasks 13.08.2020/Task 1/Program.cs
@@ -12,9 +12,9 @@ namespace Task_1
         {
             
             Console.WriteLine("Введите свой рост: ");
-            int height = Convert.ToInt32(Console.ReadLine());
-            int c = 100;
-            int weight = height - c;
+            double height = Convert.ToDouble(Console.ReadLine());
+            double c = 100;
+            double weight = height - c;
             Console.WriteLine("Ваш идеальный вес: {0}", weight );
             Console.ReadKey();
